// apps/web/src/libs/api/generated/models/resolveMctRuleParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * SaaS API
 * NestJS + Fastify backend. Auth is served by Better Auth at /api/auth/* (not part of this spec).
 * OpenAPI spec version: 1.0.0
 */
import type { ResolveMctRuleScope } from './resolveMctRuleScope';

export type ResolveMctRuleParams = {
/**
 * @pattern ^[A-Z]{3}$
 */
arrivalAirport: string;
/**
 * @pattern ^[A-Z]{3}$
 */
departureAirport: string;
scope: ResolveMctRuleScope;
/**
 * @pattern ^[A-Z0-9]{2}$
 */
arrivalAirline?: string;
/**
 * @pattern ^[A-Z0-9]{2}$
 */
departureAirline?: string;
/**
 * @minLength 1
 * @maxLength 5
 */
arrivalTerminal?: string;
/**
 * @minLength 1
 * @maxLength 5
 */
departureTerminal?: string;
};